refactor(plugin): reorganize shared module boundaries
- Move reusable rule I/O into infrastructure and split matcher, provider, server, and registry modules by responsibility.
- Decompose upgrade and RouterOS executor implementations into focused packages while preserving existing plugin facades.
- Move build capability reporting out of infra to maintain a clearer dependency direction.
- Preserve runtime behavior and validate the refactor with `cargo test --all-features` and `just check`.
