+++ src/infra/io/mod.rs
@@ -1,0 +1,6 @@
+// SPDX-FileCopyrightText: 2025 Sven Shi
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+//! Shared file and stream handling helpers.
+
+pub(crate) mod lines;
