+++ src/plugin/executor/ros_address_list/mod.rs
@@ -28,38 +28,26 @@
 //! - load persistent file-backed entries at startup and keep them fixed until
 //!   the plugin is reloaded.
 
-use std::fs;
 use std::net::IpAddr;
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::atomic::Ordering;
 use std::sync::{Arc, Mutex};
-use std::time::Duration;
 
-use ahash::AHashSet;
 use async_trait::async_trait;
-use serde::Deserialize;
-use serde_yaml_ng::Value;
 use tokio::sync::oneshot;
 use tracing::warn;
 
-use self::api::{
-    DEFAULT_CONNECT_TIMEOUT_SECS, DEFAULT_RECEIVE_TIMEOUT_SECS, DEFAULT_SEND_TIMEOUT_SECS,
-    MikrotikApi, MikrotikApiTimeouts, MikrotikRsClient,
-};
+use self::api::{MikrotikApi, MikrotikRsClient};
+use self::config::{MikrotikConfig, parse_plugin_config, validate_comment_token};
 use self::manager::{
     AddressListManager, AddressListManagerConfig, AddressListManagerHandle,
     AddressListManagerRuntime, ObserveEnqueueError,
 };
-use self::model::{AddressListFamily, AddressListKey};
+use self::metrics::RosMetrics;
 use crate::config::types::PluginConfig;
 use crate::core::context::DnsContext;
 use crate::infra::error::{DnsError, Result};
-use crate::infra::observability::metrics::{
-    MetricLabel, MetricSample, MetricSink, MetricSource, register_metric_source,
-    unregister_metric_source,
-};
-use crate::infra::system::deserialize_duration_option;
+use crate::infra::observability::metrics::{register_metric_source, unregister_metric_source};
 use crate::plugin::executor::routeros::throttle::ErrorLogThrottle;
-use crate::plugin::executor::routeros::transport::{RouterOsConnectionConfig, RouterOsTlsArgs};
 use crate::plugin::executor::routeros::{ObservedAddr, SHUTDOWN_TIMEOUT, collect_observed_addrs};
 use crate::plugin::executor::{ExecStep, Executor, ExecutorNext};
 use crate::plugin::{Plugin, PluginFactory, UninitializedPlugin};
@@ -67,353 +55,11 @@ use crate::proto::{Rcode, RecordType};
 use crate::{continue_next, plugin_factory};
 
 mod api;
+mod config;
 mod manager;
+mod metrics;
 mod model;
-
-/// Default lower TTL clamp for dynamic address-list entries.
-const DEFAULT_MIN_TTL: u32 = 60;
-/// Default upper TTL clamp for dynamic address-list entries.
-const DEFAULT_MAX_TTL: u32 = 3600;
-/// Default execution mode keeps RouterOS writes off the DNS request path.
-const DEFAULT_ASYNC_MODE: bool = true;
-/// Default shutdown behavior removes plugin-owned RouterOS entries.
-const DEFAULT_CLEANUP_ON_SHUTDOWN: bool = true;
-/// Default comment prefix used to mark OxiDNS-owned RouterOS rows.
-const DEFAULT_COMMENT_PREFIX: &str = "oxi";
-const DEFAULT_WAIT_TIMEOUT: Duration = Duration::from_secs(8);
-const DEFAULT_QUEUE_CAPACITY: usize = 16_384;
-
-#[derive(Debug, Clone, Deserialize, Default)]
-#[serde(deny_unknown_fields)]
-struct MikrotikConfigArgs {
-    /// RouterOS API endpoint, usually `<host>:8728`.
-    address: Option<String>,
-    /// RouterOS login username.
-    username: Option<String>,
-    /// RouterOS login password.
-    password: Option<String>,
-    /// RouterOS API connection timeout in seconds.
-    connect_timeout: Option<u64>,
-    /// RouterOS API command send timeout in seconds.
-    send_timeout: Option<u64>,
-    /// RouterOS API response receive timeout in seconds.
-    receive_timeout: Option<u64>,
-    /// Optional RouterOS API-SSL configuration. Presence enables TLS.
-    tls: Option<RouterOsTlsArgs>,
-    /// Whether post stage waits RouterOS writes (`false`) or queues work
-    /// (`true`).
-    #[serde(rename = "async")]
-    async_mode: Option<bool>,
-    /// Maximum time synchronous mode waits for manager completion.
-    #[serde(default, deserialize_with = "deserialize_duration_option")]
-    wait_timeout: Option<Duration>,
-    /// Maximum number of distinct keys in each manager queue stage.
-    queue_capacity: Option<usize>,
-    /// IPv4 address-list name for observed IPv4 answers.
-    address_list4: Option<String>,
-    /// IPv6 address-list name for observed IPv6 answers.
-    address_list6: Option<String>,
-    /// Prefix used in RouterOS comments to mark OxiDNS-managed entries.
-    /// Defaults to `oxi` when omitted.
-    comment_prefix: Option<String>,
-    /// Always-present address-list items that should never expire.
-    persistent: Option<PersistentArgs>,
-    /// Minimum effective TTL clamp (seconds) for observed records.
-    min_ttl: Option<u32>,
-    /// Maximum effective TTL clamp (seconds) for observed records.
-    max_ttl: Option<u32>,
-    /// Optional fixed TTL override (seconds) for dynamic observed records.
-    /// `0` means do not set RouterOS timeout.
-    fixed_ttl: Option<u32>,
-    /// Whether to clean managed address-list entries on shutdown.
-    cleanup_on_shutdown: Option<bool>,
-}
-
-#[derive(Debug, Clone, Deserialize, Default)]
-#[serde(deny_unknown_fields)]
-struct PersistentArgs {
-    /// Inline always-present IPs/CIDRs. Plain IP is normalized to host entry.
-    ips: Option<Vec<String>>,
-    /// File list that provides always-present IPs/CIDRs.
-    files: Option<Vec<String>>,
-}
-
-#[derive(Debug, Clone)]
-struct MikrotikConfig {
-    /// Connection settings consumed when the API transport is constructed.
-    connection: Option<RouterOsConnectionConfig>,
-    /// Async mode switch for post stage writes.
-    async_mode: bool,
-    wait_timeout: Duration,
-    queue_capacity: usize,
-    /// IPv4 address-list name managed by this plugin.
-    address_list4: Option<String>,
-    /// IPv6 address-list name managed by this plugin.
-    address_list6: Option<String>,
-    /// Full persistent desired set after merging inline and file sources.
-    persistent_items: AHashSet<AddressListKey>,
-    /// Prefix used in RouterOS comments to mark plugin ownership.
-    comment_prefix: String,
-    /// Minimum TTL clamp for dynamic entries.
-    min_ttl: u32,
-    /// Maximum TTL clamp for dynamic entries.
-    max_ttl: u32,
-    /// Optional fixed TTL override for dynamic entries.
-    /// `0` means do not set RouterOS timeout.
-    fixed_ttl: Option<u32>,
-    /// Whether shutdown should remove owned entries from RouterOS.
-    cleanup_on_shutdown: bool,
-}
-
-impl MikrotikConfigArgs {
-    /// Validate user-facing config and normalize it into a runtime-ready form.
-    ///
-    /// This is also where persistent input sources are parsed into normalized
-    /// `AddressListKey` values so the manager does not need to re-interpret
-    /// human-facing YAML at runtime.
-    fn into_config(self, emit_warnings: bool) -> Result<MikrotikConfig> {
-        let address = required_non_empty(self.address, "address")?;
-        let username = required_non_empty(self.username, "username")?;
-        let password = required_non_empty(self.password, "password")?;
-        let api_timeouts = MikrotikApiTimeouts::from_secs(
-            timeout_secs(
-                self.connect_timeout,
-                "connect_timeout",
-                DEFAULT_CONNECT_TIMEOUT_SECS,
-            )?,
-            timeout_secs(self.send_timeout, "send_timeout", DEFAULT_SEND_TIMEOUT_SECS)?,
-            timeout_secs(
-                self.receive_timeout,
-                "receive_timeout",
-                DEFAULT_RECEIVE_TIMEOUT_SECS,
-            )?,
-        );
-        let connection = RouterOsConnectionConfig::new(
-            address.clone(),
-            username,
-            password,
-            api_timeouts,
-            self.tls,
-        )?;
-        let address_list4 = optional_non_empty(self.address_list4);
-        let address_list6 = optional_non_empty(self.address_list6);
-        if address_list4.is_none() && address_list6.is_none() {
-            return Err(DnsError::plugin(
-                "ros_address_list requires at least one of address_list4 or address_list6",
-            ));
-        }
-
-        let comment_prefix = optional_non_empty(self.comment_prefix)
-            .unwrap_or_else(|| DEFAULT_COMMENT_PREFIX.to_string());
-        validate_comment_token("comment_prefix", &comment_prefix)?;
-
-        let min_ttl = self.min_ttl.unwrap_or(DEFAULT_MIN_TTL);
-        let max_ttl = self.max_ttl.unwrap_or(DEFAULT_MAX_TTL);
-        if min_ttl > max_ttl {
-            return Err(DnsError::plugin(format!(
-                "ros_address_list ttl range is invalid: min_ttl({min_ttl}) > max_ttl({max_ttl})"
-            )));
-        }
-        let fixed_ttl = self.fixed_ttl;
-        let wait_timeout = positive_duration(
-            self.wait_timeout.unwrap_or(DEFAULT_WAIT_TIMEOUT),
-            "wait_timeout",
-        )?;
-        let queue_capacity = positive_usize(
-            self.queue_capacity.unwrap_or(DEFAULT_QUEUE_CAPACITY),
-            "queue_capacity",
-        )?;
-
-        let parsed_persistent = parse_persistent_items(
-            self.persistent,
-            address_list4.as_deref(),
-            address_list6.as_deref(),
-        )?;
-        if emit_warnings && parsed_persistent.ignored_by_family > 0 {
-            warn!(
-                ignored = parsed_persistent.ignored_by_family,
-                "ros_address_list persistent ignored entries without corresponding address list family"
-            );
-        }
-
-        Ok(MikrotikConfig {
-            connection: Some(connection),
-            async_mode: self.async_mode.unwrap_or(DEFAULT_ASYNC_MODE),
-            wait_timeout,
-            queue_capacity,
-            address_list4,
-            address_list6,
-            persistent_items: parsed_persistent.all_items,
-            comment_prefix,
-            min_ttl,
-            max_ttl,
-            fixed_ttl,
-            cleanup_on_shutdown: self
-                .cleanup_on_shutdown
-                .unwrap_or(DEFAULT_CLEANUP_ON_SHUTDOWN),
-        })
-    }
-}
-
-#[derive(Debug)]
-struct RosMetrics {
-    tag: String,
-    observe_total: AtomicU64,
-    dropped_total: AtomicU64,
-    sync_error_total: AtomicU64,
-    sync_timeout_total: AtomicU64,
-    write_success_total: AtomicU64,
-    write_error_total: AtomicU64,
-    last_write_success_timestamp_seconds: AtomicU64,
-    pending_observations: AtomicU64,
-    managed_entries: AtomicU64,
-    coalesced_total: AtomicU64,
-    reconnect_total: AtomicU64,
-    connect_attempt_total: AtomicU64,
-    backoff_total: AtomicU64,
-    reconcile_error_total: AtomicU64,
-    last_reconcile_success_timestamp_seconds: AtomicU64,
-    degraded: AtomicU64,
-    cleanup_error_total: AtomicU64,
-}
-
-impl RosMetrics {
-    fn new(tag: String) -> Self {
-        Self {
-            tag,
-            observe_total: AtomicU64::new(0),
-            dropped_total: AtomicU64::new(0),
-            sync_error_total: AtomicU64::new(0),
-            sync_timeout_total: AtomicU64::new(0),
-            write_success_total: AtomicU64::new(0),
-            write_error_total: AtomicU64::new(0),
-            last_write_success_timestamp_seconds: AtomicU64::new(0),
-            pending_observations: AtomicU64::new(0),
-            managed_entries: AtomicU64::new(0),
-            coalesced_total: AtomicU64::new(0),
-            reconnect_total: AtomicU64::new(0),
-            connect_attempt_total: AtomicU64::new(0),
-            backoff_total: AtomicU64::new(0),
-            reconcile_error_total: AtomicU64::new(0),
-            last_reconcile_success_timestamp_seconds: AtomicU64::new(0),
-            degraded: AtomicU64::new(0),
-            cleanup_error_total: AtomicU64::new(0),
-        }
-    }
-}
-
-impl MetricSource for RosMetrics {
-    fn tag(&self) -> &str {
-        &self.tag
-    }
-
-    fn plugin_type(&self) -> &'static str {
-        "ros_address_list"
-    }
-
-    fn collect(&self, sink: &mut dyn MetricSink) {
-        let labels = [MetricLabel::new("plugin_tag", self.tag.as_str())];
-        sink.emit(MetricSample::counter(
-            "ros_address_list_observe_total",
-            "Total address observations submitted to the RouterOS address-list manager.",
-            &labels,
-            self.observe_total.load(Ordering::Relaxed),
-        ));
-        sink.emit(MetricSample::counter(
-            "ros_address_list_dropped_total",
-            "Total observations dropped in async mode (queue full or channel closed).",
-            &labels,
-            self.dropped_total.load(Ordering::Relaxed),
-        ));
-        sink.emit(MetricSample::counter(
-            "ros_address_list_sync_error_total",
-            "Total sync-mode observations that failed at the RouterOS manager.",
-            &labels,
-            self.sync_error_total.load(Ordering::Relaxed),
-        ));
-        sink.emit(MetricSample::counter(
-            "ros_address_list_sync_timeout_total",
-            "Total sync-mode observations that timed out waiting for manager completion.",
-            &labels,
-            self.sync_timeout_total.load(Ordering::Relaxed),
-        ));
-        for (name, help, value) in [
-            (
-                "ros_address_list_pending_observations",
-                "Current coalesced address-list observations waiting for processing.",
-                self.pending_observations.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_address_list_managed_entries",
-                "Current address-list entries retained by the manager.",
-                self.managed_entries.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_address_list_last_write_success_timestamp_seconds",
-                "Unix timestamp of the last successful address-list upsert.",
-                self.last_write_success_timestamp_seconds
-                    .load(Ordering::Relaxed),
-            ),
-            (
-                "ros_address_list_last_reconcile_success_timestamp_seconds",
-                "Unix timestamp of the last successful address-list reconcile.",
-                self.last_reconcile_success_timestamp_seconds
-                    .load(Ordering::Relaxed),
-            ),
-            (
-                "ros_address_list_degraded",
-                "Whether the RouterOS transport is currently degraded.",
-                self.degraded.load(Ordering::Relaxed),
-            ),
-        ] {
-            sink.emit(MetricSample::gauge(name, help, &labels, value));
-        }
-        for (name, help, value) in [
-            (
-                "ros_address_list_write_success_total",
-                "Total successful RouterOS address-list upserts.",
-                self.write_success_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_address_list_write_error_total",
-                "Total failed RouterOS address-list upserts.",
-                self.write_error_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_address_list_coalesced_total",
-                "Total address-list observations merged into an existing mailbox key.",
-                self.coalesced_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_address_list_reconnect_total",
-                "Total successful RouterOS transport reconnections.",
-                self.reconnect_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_address_list_connect_attempt_total",
-                "Total RouterOS transport connection attempts.",
-                self.connect_attempt_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_address_list_backoff_total",
-                "Total RouterOS transport backoff schedules.",
-                self.backoff_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_address_list_reconcile_error_total",
-                "Total failed address-list reconcile attempts.",
-                self.reconcile_error_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_address_list_cleanup_error_total",
-                "Total address-list entries that failed shutdown cleanup.",
-                self.cleanup_error_total.load(Ordering::Relaxed),
-            ),
-        ] {
-            sink.emit(MetricSample::counter(name, help, &labels, value));
-        }
-    }
-}
+mod persistent;
 
 #[derive(Debug)]
 struct MikrotikExecutor {
@@ -657,277 +303,25 @@ fn extract_observation(
     Some(addrs)
 }
 
-fn parse_plugin_config(args: Option<Value>, emit_warnings: bool) -> Result<MikrotikConfig> {
-    let Some(args) = args else {
-        return Err(DnsError::plugin("ros_address_list plugin requires args"));
-    };
-    let raw = serde_yaml_ng::from_value::<MikrotikConfigArgs>(args)
-        .map_err(|e| DnsError::plugin(format!("failed to parse ros_address_list config: {e}")))?;
-    raw.into_config(emit_warnings)
-}
-
-fn required_non_empty(value: Option<String>, field: &str) -> Result<String> {
-    let Some(value) = value else {
-        return Err(DnsError::plugin(format!(
-            "ros_address_list '{field}' is required"
-        )));
-    };
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        return Err(DnsError::plugin(format!(
-            "ros_address_list '{field}' cannot be empty"
-        )));
-    }
-    Ok(trimmed.to_string())
-}
-
-fn optional_non_empty(value: Option<String>) -> Option<String> {
-    value
-        .map(|v| v.trim().to_string())
-        .filter(|v| !v.is_empty())
-}
-
-fn timeout_secs(value: Option<u64>, field: &str, default_secs: u64) -> Result<u64> {
-    match value {
-        Some(0) => Err(DnsError::plugin(format!(
-            "ros_address_list '{field}' must be greater than 0 seconds"
-        ))),
-        Some(value) => Ok(value),
-        None => Ok(default_secs),
-    }
-}
-
-fn positive_duration(value: Duration, field: &str) -> Result<Duration> {
-    if value.is_zero() {
-        return Err(DnsError::plugin(format!(
-            "ros_address_list '{field}' must be greater than 0"
-        )));
-    }
-    Ok(value)
-}
-
-fn positive_usize(value: usize, field: &str) -> Result<usize> {
-    if value == 0 {
-        return Err(DnsError::plugin(format!(
-            "ros_address_list '{field}' must be greater than 0"
-        )));
-    }
-    Ok(value)
-}
-
-#[inline]
-fn contains_comment_delimiter(value: &str) -> bool {
-    value.contains(';') || value.contains('=')
-}
-
-fn validate_comment_token(field: &str, value: &str) -> Result<()> {
-    if contains_comment_delimiter(value) {
-        return Err(DnsError::plugin(format!(
-            "ros_address_list '{field}' cannot contain ';' or '='"
-        )));
-    }
-    Ok(())
-}
-
-#[derive(Debug, Default)]
-struct ParsedPersistentItems {
-    /// Final desired set after merging inline and file sources.
-    all_items: AHashSet<AddressListKey>,
-    /// Count of items skipped because that family is not configured.
-    ignored_by_family: usize,
-}
-
-/// Parse `persistent` config into normalized address-list keys.
-///
-/// The parser performs all expensive normalization and validation at startup:
-/// plain IPs become host prefixes, CIDRs are masked to network form, and each
-/// item is bound to the correct IPv4/IPv6 address-list name.
-fn parse_persistent_items(
-    persistent: Option<PersistentArgs>,
-    address_list4: Option<&str>,
-    address_list6: Option<&str>,
-) -> Result<ParsedPersistentItems> {
-    let mut parsed = ParsedPersistentItems::default();
-    let Some(persistent) = persistent else {
-        return Ok(parsed);
-    };
-
-    if let Some(ips) = persistent.ips {
-        for (index, item) in ips.into_iter().enumerate() {
-            let source = format!("persistent.ips[{index}]");
-            let key = parse_persistent_item(
-                item.as_str(),
-                source.as_str(),
-                address_list4,
-                address_list6,
-            )?;
-            match key {
-                Some(key) => {
-                    parsed.all_items.insert(key);
-                }
-                None => {
-                    parsed.ignored_by_family = parsed.ignored_by_family.saturating_add(1);
-                }
-            }
-        }
-    }
-
-    let files = parse_persistent_files(persistent.files)?;
-    let (file_items, ignored_by_family) =
-        load_persistent_items_from_files(files.as_slice(), address_list4, address_list6)?;
-    parsed.ignored_by_family = parsed.ignored_by_family.saturating_add(ignored_by_family);
-    parsed.all_items.extend(file_items);
-    Ok(parsed)
-}
-
-fn parse_persistent_files(files: Option<Vec<String>>) -> Result<Vec<String>> {
-    let mut out = Vec::new();
-    let Some(files) = files else {
-        return Ok(out);
-    };
-    for (index, file_raw) in files.into_iter().enumerate() {
-        let file = file_raw.trim();
-        if file.is_empty() {
-            return Err(DnsError::plugin(format!(
-                "ros_address_list persistent.files[{index}] cannot be empty"
-            )));
-        }
-        out.push(file.to_string());
-    }
-    Ok(out)
-}
-
-/// Parse one file body into normalized persistent items.
-///
-/// Files use the same item grammar as inline YAML. Empty lines and `#` comments
-/// are ignored. Family-mismatched entries are skipped rather than failing
-/// startup so shared source files can contain both IPv4 and IPv6 items.
-fn load_persistent_items_from_content(
-    source_prefix: &str,
-    content: &str,
-    address_list4: Option<&str>,
-    address_list6: Option<&str>,
-) -> Result<(AHashSet<AddressListKey>, usize)> {
-    let mut out = AHashSet::new();
-    let mut ignored_by_family = 0usize;
-
-    for (line_no, line) in content.lines().enumerate() {
-        let token = line.split('#').next().unwrap_or_default().trim();
-        if token.is_empty() {
-            continue;
-        }
-
-        let source = format!("{source_prefix} line {}", line_no + 1);
-        match parse_persistent_item(token, source.as_str(), address_list4, address_list6)? {
-            Some(key) => {
-                out.insert(key);
-            }
-            None => {
-                ignored_by_family = ignored_by_family.saturating_add(1);
-            }
-        }
-    }
-
-    Ok((out, ignored_by_family))
-}
-
-fn load_persistent_items_from_files(
-    files: &[String],
-    address_list4: Option<&str>,
-    address_list6: Option<&str>,
-) -> Result<(AHashSet<AddressListKey>, usize)> {
-    let mut out = AHashSet::new();
-    let mut ignored_by_family = 0usize;
-
-    for (index, file) in files.iter().enumerate() {
-        let content = fs::read_to_string(file).map_err(|e| {
-            DnsError::plugin(format!(
-                "ros_address_list failed to read persistent file '{file}': {e}"
-            ))
-        })?;
-        let source_prefix = format!("persistent.files[{index}]");
-        let (loaded, ignored_delta) = load_persistent_items_from_content(
-            source_prefix.as_str(),
-            &content,
-            address_list4,
-            address_list6,
-        )?;
-        out.extend(loaded);
-        ignored_by_family = ignored_by_family.saturating_add(ignored_delta);
-    }
-
-    Ok((out, ignored_by_family))
-}
-
-/// Parse one human-facing persistent item and bind it to the correct list.
-///
-/// Return `Ok(None)` when the item is valid but its IP family has no configured
-/// target list, allowing callers to ignore mixed-family source files cleanly.
-fn parse_persistent_item(
-    raw: &str,
-    source: &str,
-    address_list4: Option<&str>,
-    address_list6: Option<&str>,
-) -> Result<Option<AddressListKey>> {
-    let value = raw.trim();
-    if value.is_empty() {
-        return Err(DnsError::plugin(format!(
-            "ros_address_list {source} is empty"
-        )));
-    }
-
-    let (ip, prefix) = if let Some((ip_raw, prefix_raw)) = value.split_once('/') {
-        let ip = ip_raw.trim().parse::<IpAddr>().map_err(|e| {
-            DnsError::plugin(format!(
-                "ros_address_list {source} has invalid ip '{ip_raw}': {e}"
-            ))
-        })?;
-        let prefix = prefix_raw.trim().parse::<u8>().map_err(|e| {
-            DnsError::plugin(format!(
-                "ros_address_list {source} has invalid prefix '{prefix_raw}': {e}"
-            ))
-        })?;
-        (ip, prefix)
-    } else {
-        let ip = value.parse::<IpAddr>().map_err(|e| {
-            DnsError::plugin(format!(
-                "ros_address_list {source} has invalid ip '{value}': {e}"
-            ))
-        })?;
-        let family = AddressListFamily::from_ip(ip);
-        (ip, family.host_prefix())
-    };
-
-    let family = AddressListFamily::from_ip(ip);
-    let list = match family {
-        AddressListFamily::Ipv4 => address_list4,
-        AddressListFamily::Ipv6 => address_list6,
-    };
-    let Some(list) = list else {
-        return Ok(None);
-    };
-
-    AddressListKey::new_with_prefix(ip, prefix, list.to_string())
-        .ok_or_else(|| {
-            DnsError::plugin(format!(
-                "ros_address_list {source} has invalid prefix /{prefix} for {ip}"
-            ))
-        })
-        .map(Some)
-}
-
 #[cfg(test)]
 mod tests {
     use std::net::{Ipv4Addr, Ipv6Addr, SocketAddr};
     use std::sync::atomic::AtomicUsize;
+    use std::time::Duration;
 
-    use ahash::AHashMap;
+    use ahash::{AHashMap, AHashSet};
+    use serde_yaml_ng::Value;
 
+    use super::config::{
+        DEFAULT_COMMENT_PREFIX, DEFAULT_MAX_TTL, DEFAULT_MIN_TTL, DEFAULT_QUEUE_CAPACITY,
+        DEFAULT_WAIT_TIMEOUT,
+    };
+    use super::persistent::{load_persistent_items_from_content, parse_persistent_files};
     use super::*;
     use crate::infra::clock::AppClock;
-    use crate::plugin::executor::ros_address_list::api::RouterListEntry;
+    use crate::plugin::executor::ros_address_list::api::{MikrotikApiTimeouts, RouterListEntry};
     use crate::plugin::executor::ros_address_list::model::{
-        OwnedCommentKind, decode_owned_comment, encode_comment,
+        AddressListFamily, AddressListKey, OwnedCommentKind, decode_owned_comment, encode_comment,
     };
     use crate::proto::rdata::{A, AAAA};
     use crate::proto::{DNSClass, Message, Name, Question, RData, Rcode, Record, RecordType};
