+++ src/plugin/executor/ros_route/mod.rs
@@ -29,249 +29,44 @@
 //!   optional shutdown cleanup.
 //! - assume routing table/rule/default routes are already provisioned by users.
 
-use std::fs;
-use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::net::IpAddr;
+use std::sync::atomic::Ordering;
 use std::sync::{Arc, Mutex};
-use std::time::Duration;
 
 use ahash::AHashSet;
 use async_trait::async_trait;
-use serde::Deserialize;
-use serde_yaml_ng::Value;
 use tokio::sync::oneshot;
 use tracing::warn;
 
+use self::api::{MikrotikApi, MikrotikRsClient};
+use self::config::{MikrotikConfig, parse_plugin_config, validate_comment_token};
+use self::manager::{
+    ObserveEnqueueError, RouteManager, RouteManagerConfig, RouteManagerHandle, RouteManagerRuntime,
+};
+use self::metrics::RosRouteMetrics;
 use crate::config::types::PluginConfig;
 use crate::core::context::DnsContext;
 use crate::infra::error::{DnsError, Result};
-use crate::infra::observability::metrics::{
-    MetricLabel, MetricSample, MetricSink, MetricSource, register_metric_source,
-    unregister_metric_source,
-};
+use crate::infra::observability::metrics::{register_metric_source, unregister_metric_source};
+use crate::plugin::executor::routeros::ip_prefix::IpPrefix;
+use crate::plugin::executor::routeros::throttle::ErrorLogThrottle;
+use crate::plugin::executor::routeros::{ObservedAddr, SHUTDOWN_TIMEOUT, collect_observed_addrs};
 use crate::plugin::executor::{ExecStep, Executor, ExecutorNext};
 use crate::plugin::{Plugin, PluginFactory, UninitializedPlugin};
 use crate::proto::{Rcode, RecordType};
 use crate::{continue_next, plugin_factory};
 
-const DEFAULT_MIN_TTL: u32 = 60;
-const DEFAULT_MAX_TTL: u32 = 3600;
-const DEFAULT_ASYNC_MODE: bool = true;
-const DEFAULT_CLEANUP_ON_SHUTDOWN: bool = true;
-const DEFAULT_CONNTRACK_GUARD: bool = false;
-const DEFAULT_ROUTE_DISTANCE: u8 = 100;
-const DEFAULT_COMMENT_PREFIX: &str = "oxi";
-const DEFAULT_WAIT_TIMEOUT: Duration = Duration::from_secs(8);
-const DEFAULT_QUEUE_CAPACITY: usize = 16_384;
-
-#[derive(Debug, Clone, Deserialize, Default)]
-#[serde(deny_unknown_fields)]
-struct MikrotikConfigArgs {
-    /// RouterOS API endpoint, usually `<host>:8728`.
-    address: Option<String>,
-    /// RouterOS login username.
-    username: Option<String>,
-    /// RouterOS login password.
-    password: Option<String>,
-    /// RouterOS API connection timeout in seconds.
-    connect_timeout: Option<u64>,
-    /// RouterOS API command send timeout in seconds.
-    send_timeout: Option<u64>,
-    /// RouterOS API response receive timeout in seconds.
-    receive_timeout: Option<u64>,
-    /// Optional RouterOS API-SSL configuration. Presence enables TLS.
-    tls: Option<RouterOsTlsArgs>,
-    /// Whether post stage waits RouterOS writes (`false`) or queues work
-    /// (`true`).
-    #[serde(rename = "async")]
-    async_mode: Option<bool>,
-    #[serde(default, deserialize_with = "deserialize_duration_option")]
-    wait_timeout: Option<Duration>,
-    queue_capacity: Option<usize>,
-    /// Dedicated RouterOS routing table for managed routes.
-    routing_table: Option<String>,
-    /// IPv4 gateway value for managed IPv4 routes.
-    gateway4: Option<String>,
-    /// IPv6 gateway value for managed IPv6 routes.
-    gateway6: Option<String>,
-    /// Prefix used in RouterOS route comments to mark OxiDNS-managed routes.
-    /// Defaults to `oxi` when omitted.
-    comment_prefix: Option<String>,
-    /// Route distance written to RouterOS for managed routes.
-    distance: Option<u8>,
-    /// Always-present routes that should not expire with DNS TTL.
-    persistent: Option<PersistentArgs>,
-    /// Minimum effective TTL clamp (seconds) for observed records.
-    min_ttl: Option<u32>,
-    /// Maximum effective TTL clamp (seconds) for observed records.
-    max_ttl: Option<u32>,
-    /// Optional fixed TTL override (seconds) for dynamic observed records.
-    /// `0` keeps a dynamic route until explicit cleanup or operator removal.
-    fixed_ttl: Option<u32>,
-    /// Whether to clean managed dynamic routes on shutdown.
-    cleanup_on_shutdown: Option<bool>,
-    /// Delay normal route removal while RouterOS connection tracking has a
-    /// connection for the route destination.
-    conntrack_guard: Option<bool>,
-}
-
-#[derive(Debug, Clone, Deserialize, Default)]
-#[serde(deny_unknown_fields)]
-struct PersistentArgs {
-    /// Inline always-present IPs/CIDRs. Plain IP is normalized to host route.
-    ips: Option<Vec<String>>,
-    /// File list that provides always-present IPs.
-    files: Option<Vec<String>>,
-}
-
-#[derive(Debug, Clone)]
-struct MikrotikConfig {
-    /// Connection settings consumed when the API transport is constructed.
-    connection: Option<RouterOsConnectionConfig>,
-    /// Async mode switch for post stage RouterOS writes.
-    async_mode: bool,
-    wait_timeout: Duration,
-    queue_capacity: usize,
-    /// Dedicated RouterOS routing table for this plugin.
-    routing_table: String,
-    /// Optional IPv4 gateway.
-    gateway4: Option<String>,
-    /// Optional IPv6 gateway.
-    gateway6: Option<String>,
-    /// Always-present routes in normalized CIDR format (`ip/prefix`).
-    persistent_ips: AHashSet<String>,
-    /// Managed route comment prefix.
-    comment_prefix: String,
-    /// Route distance written to RouterOS.
-    distance: u8,
-    /// Minimum effective TTL clamp in seconds.
-    min_ttl: u32,
-    /// Maximum effective TTL clamp in seconds.
-    max_ttl: u32,
-    /// Optional fixed TTL override in seconds. `0` never expires by time.
-    fixed_ttl: Option<u32>,
-    /// Shutdown cleanup behavior for dynamic routes.
-    cleanup_on_shutdown: bool,
-    /// Delay normal route removal while a matching RouterOS connection exists.
-    conntrack_guard: bool,
-}
+mod api;
+mod config;
+mod manager;
+mod metrics;
+mod model;
+mod persistent;
 
 #[derive(Debug)]
 struct ExtractedObservation {
     addrs: Vec<ObservedAddr>,
 }
-
-impl MikrotikConfigArgs {
-    fn into_config(self, emit_warnings: bool) -> Result<MikrotikConfig> {
-        let address = required_non_empty(self.address, "address")?;
-        let username = required_non_empty(self.username, "username")?;
-        let password = required_non_empty(self.password, "password")?;
-        let api_timeouts = MikrotikApiTimeouts::from_secs(
-            timeout_secs(
-                self.connect_timeout,
-                "connect_timeout",
-                DEFAULT_CONNECT_TIMEOUT_SECS,
-            )?,
-            timeout_secs(self.send_timeout, "send_timeout", DEFAULT_SEND_TIMEOUT_SECS)?,
-            timeout_secs(
-                self.receive_timeout,
-                "receive_timeout",
-                DEFAULT_RECEIVE_TIMEOUT_SECS,
-            )?,
-        );
-        let connection = RouterOsConnectionConfig::new(
-            address.clone(),
-            username,
-            password,
-            api_timeouts,
-            self.tls,
-        )?;
-        let routing_table = required_non_empty(self.routing_table, "routing_table")?;
-        let comment_prefix = optional_non_empty(self.comment_prefix)
-            .unwrap_or_else(|| DEFAULT_COMMENT_PREFIX.to_string());
-        validate_comment_token("comment_prefix", &comment_prefix)?;
-        let distance = self.distance.unwrap_or(DEFAULT_ROUTE_DISTANCE);
-
-        let gateway4 = optional_non_empty(self.gateway4);
-        let gateway6 = optional_non_empty(self.gateway6);
-        if gateway4.is_none() && gateway6.is_none() {
-            return Err(DnsError::plugin(
-                "ros_route requires at least one of gateway4 or gateway6",
-            ));
-        }
-
-        let min_ttl = self.min_ttl.unwrap_or(DEFAULT_MIN_TTL);
-        let max_ttl = self.max_ttl.unwrap_or(DEFAULT_MAX_TTL);
-        if min_ttl > max_ttl {
-            return Err(DnsError::plugin(format!(
-                "ros_route ttl range is invalid: min_ttl({min_ttl}) > max_ttl({max_ttl})"
-            )));
-        }
-        // `0` deliberately means a dynamic route that never expires by time.
-        let fixed_ttl = self.fixed_ttl;
-        let wait_timeout = positive_duration(
-            self.wait_timeout.unwrap_or(DEFAULT_WAIT_TIMEOUT),
-            "wait_timeout",
-        )?;
-        let queue_capacity = positive_usize(
-            self.queue_capacity.unwrap_or(DEFAULT_QUEUE_CAPACITY),
-            "queue_capacity",
-        )?;
-        let parsed_persistent =
-            parse_persistent_ips(self.persistent, gateway4.is_some(), gateway6.is_some())?;
-        let ignored_by_gateway = parsed_persistent.ignored_by_gateway;
-        if emit_warnings && ignored_by_gateway > 0 {
-            warn!(
-                ignored = ignored_by_gateway,
-                "ros_route persistent ignored entries without corresponding gateway family"
-            );
-        }
-        let ignored_default_route = parsed_persistent.ignored_default_route;
-        if emit_warnings && ignored_default_route > 0 {
-            warn!(
-                ignored = ignored_default_route,
-                "ros_route persistent ignored default-route entries (/0)"
-            );
-        }
-
-        Ok(MikrotikConfig {
-            connection: Some(connection),
-            async_mode: self.async_mode.unwrap_or(DEFAULT_ASYNC_MODE),
-            wait_timeout,
-            queue_capacity,
-            routing_table,
-            gateway4,
-            gateway6,
-            persistent_ips: parsed_persistent.all_ips,
-            comment_prefix,
-            distance,
-            min_ttl,
-            max_ttl,
-            fixed_ttl,
-            cleanup_on_shutdown: self
-                .cleanup_on_shutdown
-                .unwrap_or(DEFAULT_CLEANUP_ON_SHUTDOWN),
-            conntrack_guard: self.conntrack_guard.unwrap_or(DEFAULT_CONNTRACK_GUARD),
-        })
-    }
-}
-
-mod api;
-mod manager;
-mod model;
-
-use self::api::{
-    DEFAULT_CONNECT_TIMEOUT_SECS, DEFAULT_RECEIVE_TIMEOUT_SECS, DEFAULT_SEND_TIMEOUT_SECS,
-    MikrotikApi, MikrotikApiTimeouts, MikrotikRsClient,
-};
-use self::manager::{
-    ObserveEnqueueError, RouteManager, RouteManagerConfig, RouteManagerHandle, RouteManagerRuntime,
-};
-use crate::infra::system::deserialize_duration_option;
-use crate::plugin::executor::routeros::ip_prefix::IpPrefix;
-use crate::plugin::executor::routeros::throttle::ErrorLogThrottle;
-use crate::plugin::executor::routeros::transport::{RouterOsConnectionConfig, RouterOsTlsArgs};
-use crate::plugin::executor::routeros::{ObservedAddr, SHUTDOWN_TIMEOUT, collect_observed_addrs};
 
 #[derive(Debug)]
 struct MikrotikExecutor {
@@ -282,182 +77,6 @@ struct MikrotikExecutor {
     manager_handle: Option<RouteManagerHandle>,
     runtime: Mutex<Option<RouteManagerRuntime>>,
     queue_logs: ErrorLogThrottle,
-}
-
-#[derive(Debug)]
-struct RosRouteMetrics {
-    tag: String,
-    observe_total: AtomicU64,
-    dropped_total: AtomicU64,
-    sync_error_total: AtomicU64,
-    sync_timeout_total: AtomicU64,
-    write_success_total: AtomicU64,
-    write_error_total: AtomicU64,
-    last_write_success_timestamp_seconds: AtomicU64,
-    delete_deferred_total: AtomicU64,
-    connection_check_error_total: AtomicU64,
-    pending_observations: AtomicU64,
-    managed_entries: AtomicU64,
-    coalesced_total: AtomicU64,
-    reconnect_total: AtomicU64,
-    connect_attempt_total: AtomicU64,
-    backoff_total: AtomicU64,
-    reconcile_error_total: AtomicU64,
-    last_reconcile_success_timestamp_seconds: AtomicU64,
-    degraded: AtomicU64,
-    cleanup_error_total: AtomicU64,
-}
-
-impl RosRouteMetrics {
-    fn new(tag: String) -> Self {
-        Self {
-            tag,
-            observe_total: AtomicU64::new(0),
-            dropped_total: AtomicU64::new(0),
-            sync_error_total: AtomicU64::new(0),
-            sync_timeout_total: AtomicU64::new(0),
-            write_success_total: AtomicU64::new(0),
-            write_error_total: AtomicU64::new(0),
-            last_write_success_timestamp_seconds: AtomicU64::new(0),
-            delete_deferred_total: AtomicU64::new(0),
-            connection_check_error_total: AtomicU64::new(0),
-            pending_observations: AtomicU64::new(0),
-            managed_entries: AtomicU64::new(0),
-            coalesced_total: AtomicU64::new(0),
-            reconnect_total: AtomicU64::new(0),
-            connect_attempt_total: AtomicU64::new(0),
-            backoff_total: AtomicU64::new(0),
-            reconcile_error_total: AtomicU64::new(0),
-            last_reconcile_success_timestamp_seconds: AtomicU64::new(0),
-            degraded: AtomicU64::new(0),
-            cleanup_error_total: AtomicU64::new(0),
-        }
-    }
-}
-
-impl MetricSource for RosRouteMetrics {
-    fn tag(&self) -> &str {
-        &self.tag
-    }
-
-    fn plugin_type(&self) -> &'static str {
-        "ros_route"
-    }
-
-    fn collect(&self, sink: &mut dyn MetricSink) {
-        let labels = [MetricLabel::new("plugin_tag", self.tag.as_str())];
-        sink.emit(MetricSample::counter(
-            "ros_route_observe_total",
-            "Total address observations submitted to the RouterOS route manager.",
-            &labels,
-            self.observe_total.load(Ordering::Relaxed),
-        ));
-        sink.emit(MetricSample::counter(
-            "ros_route_dropped_total",
-            "Total route observations dropped because the manager queue was unavailable.",
-            &labels,
-            self.dropped_total.load(Ordering::Relaxed),
-        ));
-        sink.emit(MetricSample::counter(
-            "ros_route_sync_error_total",
-            "Total synchronous route observations that failed without changing DNS output.",
-            &labels,
-            self.sync_error_total.load(Ordering::Relaxed),
-        ));
-        sink.emit(MetricSample::counter(
-            "ros_route_sync_timeout_total",
-            "Total synchronous route observations that timed out without changing DNS output.",
-            &labels,
-            self.sync_timeout_total.load(Ordering::Relaxed),
-        ));
-        sink.emit(MetricSample::counter(
-            "ros_route_delete_deferred_total",
-            "Total route deletions deferred because a matching RouterOS connection exists.",
-            &labels,
-            self.delete_deferred_total.load(Ordering::Relaxed),
-        ));
-        sink.emit(MetricSample::counter(
-            "ros_route_connection_check_error_total",
-            "Total RouterOS connection-tracking queries that failed during route deletion.",
-            &labels,
-            self.connection_check_error_total.load(Ordering::Relaxed),
-        ));
-        for (name, help, value) in [
-            (
-                "ros_route_pending_observations",
-                "Current coalesced route observations waiting for processing.",
-                self.pending_observations.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_route_managed_entries",
-                "Current route entries retained by the manager.",
-                self.managed_entries.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_route_last_write_success_timestamp_seconds",
-                "Unix timestamp of the last successful route upsert.",
-                self.last_write_success_timestamp_seconds
-                    .load(Ordering::Relaxed),
-            ),
-            (
-                "ros_route_last_reconcile_success_timestamp_seconds",
-                "Unix timestamp of the last successful route reconcile.",
-                self.last_reconcile_success_timestamp_seconds
-                    .load(Ordering::Relaxed),
-            ),
-            (
-                "ros_route_degraded",
-                "Whether the RouterOS transport is currently degraded.",
-                self.degraded.load(Ordering::Relaxed),
-            ),
-        ] {
-            sink.emit(MetricSample::gauge(name, help, &labels, value));
-        }
-        for (name, help, value) in [
-            (
-                "ros_route_write_success_total",
-                "Total successful RouterOS route upserts.",
-                self.write_success_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_route_write_error_total",
-                "Total failed RouterOS route upserts.",
-                self.write_error_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_route_coalesced_total",
-                "Total route observations merged into an existing mailbox key.",
-                self.coalesced_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_route_reconnect_total",
-                "Total successful RouterOS transport reconnections.",
-                self.reconnect_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_route_connect_attempt_total",
-                "Total RouterOS transport connection attempts.",
-                self.connect_attempt_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_route_backoff_total",
-                "Total RouterOS transport backoff schedules.",
-                self.backoff_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_route_reconcile_error_total",
-                "Total failed route reconcile attempts.",
-                self.reconcile_error_total.load(Ordering::Relaxed),
-            ),
-            (
-                "ros_route_cleanup_error_total",
-                "Total route entries that failed shutdown cleanup.",
-                self.cleanup_error_total.load(Ordering::Relaxed),
-            ),
-        ] {
-            sink.emit(MetricSample::counter(name, help, &labels, value));
-        }
-    }
 }
 
 #[async_trait]
@@ -681,325 +300,14 @@ fn extract_observation(
     (!addrs.is_empty()).then_some(ExtractedObservation { addrs })
 }
 
-fn parse_plugin_config(args: Option<Value>, emit_warnings: bool) -> Result<MikrotikConfig> {
-    let Some(args) = args else {
-        return Err(DnsError::plugin("ros_route plugin requires args"));
-    };
-    let raw = serde_yaml_ng::from_value::<MikrotikConfigArgs>(args)
-        .map_err(|e| DnsError::plugin(format!("failed to parse ros_route config: {e}")))?;
-    raw.into_config(emit_warnings)
-}
-
-/// Require non-empty string config fields and keep trimmed value.
-fn required_non_empty(value: Option<String>, field: &str) -> Result<String> {
-    let Some(value) = value else {
-        return Err(DnsError::plugin(format!("ros_route '{field}' is required")));
-    };
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        return Err(DnsError::plugin(format!(
-            "ros_route '{field}' cannot be empty"
-        )));
-    }
-    Ok(trimmed.to_string())
-}
-
-/// Convert optional string to trimmed non-empty value.
-fn optional_non_empty(value: Option<String>) -> Option<String> {
-    value
-        .map(|v| v.trim().to_string())
-        .filter(|v| !v.is_empty())
-}
-
-#[inline]
-fn contains_comment_delimiter(value: &str) -> bool {
-    value.contains(';') || value.contains('=')
-}
-
-fn validate_comment_token(field: &str, value: &str) -> Result<()> {
-    if contains_comment_delimiter(value) {
-        return Err(DnsError::plugin(format!(
-            "ros_route '{field}' cannot contain ';' or '='"
-        )));
-    }
-    Ok(())
-}
-
-fn timeout_secs(value: Option<u64>, field: &str, default_secs: u64) -> Result<u64> {
-    match value {
-        Some(0) => Err(DnsError::plugin(format!(
-            "ros_route '{field}' must be greater than 0 seconds"
-        ))),
-        Some(value) => Ok(value),
-        None => Ok(default_secs),
-    }
-}
-
-fn positive_duration(value: Duration, field: &str) -> Result<Duration> {
-    if value.is_zero() {
-        return Err(DnsError::plugin(format!(
-            "ros_route '{field}' must be greater than 0"
-        )));
-    }
-    Ok(value)
-}
-
-fn positive_usize(value: usize, field: &str) -> Result<usize> {
-    if value == 0 {
-        return Err(DnsError::plugin(format!(
-            "ros_route '{field}' must be greater than 0"
-        )));
-    }
-    Ok(value)
-}
-
-#[derive(Debug, Default)]
-struct ParsedPersistentRoutes {
-    all_ips: AHashSet<String>,
-    inline_ips: AHashSet<String>,
-    files: Vec<String>,
-    ignored_by_gateway: usize,
-    ignored_default_route: usize,
-}
-
-/// Parse always-present route list from inline args and optional files.
-///
-/// Accepted item formats:
-/// - `1.1.1.1`
-/// - `2001:db8::1`
-/// - generic CIDR: `1.1.1.0/24`, `2001:db8::/64`
-///
-/// Entries whose IP family has no corresponding configured gateway are skipped.
-fn parse_persistent_ips(
-    persistent: Option<PersistentArgs>,
-    gateway4_enabled: bool,
-    gateway6_enabled: bool,
-) -> Result<ParsedPersistentRoutes> {
-    let mut parsed = ParsedPersistentRoutes::default();
-    let Some(route) = persistent else {
-        return Ok(parsed);
-    };
-
-    if let Some(ips) = route.ips {
-        for (index, item) in ips.into_iter().enumerate() {
-            let source = format!("persistent.ips[{index}]");
-            let cidr = parse_persistent_ip_item(item.as_str(), source.as_str())?;
-            if is_default_route_cidr(cidr.as_str()) {
-                parsed.ignored_default_route = parsed.ignored_default_route.saturating_add(1);
-                continue;
-            }
-            if !is_persistent_ip_family_enabled(
-                cidr.as_str(),
-                gateway4_enabled,
-                gateway6_enabled,
-                source.as_str(),
-            )? {
-                parsed.ignored_by_gateway = parsed.ignored_by_gateway.saturating_add(1);
-                continue;
-            }
-            parsed.inline_ips.insert(cidr.clone());
-            parsed.all_ips.insert(cidr);
-        }
-    }
-
-    parsed.files = parse_persistent_files(route.files)?;
-    let (file_ips, ignored_from_files, ignored_default_from_files) =
-        load_persistent_ips_from_files(
-            parsed.files.as_slice(),
-            gateway4_enabled,
-            gateway6_enabled,
-        )?;
-    parsed.ignored_by_gateway = parsed.ignored_by_gateway.saturating_add(ignored_from_files);
-    parsed.ignored_default_route = parsed
-        .ignored_default_route
-        .saturating_add(ignored_default_from_files);
-    parsed.all_ips.extend(file_ips);
-
-    Ok(parsed)
-}
-
-fn parse_persistent_files(files: Option<Vec<String>>) -> Result<Vec<String>> {
-    let mut out = Vec::new();
-    let Some(files) = files else {
-        return Ok(out);
-    };
-    for (index, file_raw) in files.into_iter().enumerate() {
-        let file = file_raw.trim();
-        if file.is_empty() {
-            return Err(DnsError::plugin(format!(
-                "ros_route persistent.files[{index}] cannot be empty"
-            )));
-        }
-        out.push(file.to_string());
-    }
-    Ok(out)
-}
-
-fn load_persistent_ips_from_content(
-    source_prefix: &str,
-    content: &str,
-    gateway4_enabled: bool,
-    gateway6_enabled: bool,
-) -> Result<(AHashSet<String>, usize, usize)> {
-    let mut out = AHashSet::new();
-    let mut ignored_by_gateway = 0usize;
-    let mut ignored_default_route = 0usize;
-
-    for (line_no, line) in content.lines().enumerate() {
-        let token = line.split('#').next().unwrap_or_default().trim();
-        if token.is_empty() {
-            continue;
-        }
-
-        let source = format!("{source_prefix} line {}", line_no + 1);
-        let cidr = parse_persistent_ip_item(token, source.as_str())?;
-        if is_default_route_cidr(cidr.as_str()) {
-            ignored_default_route = ignored_default_route.saturating_add(1);
-            continue;
-        }
-        if !is_persistent_ip_family_enabled(
-            cidr.as_str(),
-            gateway4_enabled,
-            gateway6_enabled,
-            source.as_str(),
-        )? {
-            ignored_by_gateway = ignored_by_gateway.saturating_add(1);
-            continue;
-        }
-        out.insert(cidr);
-    }
-
-    Ok((out, ignored_by_gateway, ignored_default_route))
-}
-
-pub(super) fn load_persistent_ips_from_files(
-    files: &[String],
-    gateway4_enabled: bool,
-    gateway6_enabled: bool,
-) -> Result<(AHashSet<String>, usize, usize)> {
-    let mut out = AHashSet::new();
-    let mut ignored_by_gateway = 0usize;
-    let mut ignored_default_route = 0usize;
-
-    for (index, file) in files.iter().enumerate() {
-        let content = fs::read_to_string(file).map_err(|e| {
-            DnsError::plugin(format!(
-                "ros_route failed to read persistent route file '{file}': {e}"
-            ))
-        })?;
-        let source_prefix = format!("persistent.files[{index}]");
-        let (loaded, ignored_by_gateway_delta, ignored_default_delta) =
-            load_persistent_ips_from_content(
-                source_prefix.as_str(),
-                &content,
-                gateway4_enabled,
-                gateway6_enabled,
-            )?;
-        out.extend(loaded);
-        ignored_by_gateway = ignored_by_gateway.saturating_add(ignored_by_gateway_delta);
-        ignored_default_route = ignored_default_route.saturating_add(ignored_default_delta);
-    }
-
-    Ok((out, ignored_by_gateway, ignored_default_route))
-}
-
-/// Parse one persistent item and normalize into `ip/prefix`.
-///
-/// Rules:
-/// - plain IPv4/IPv6 becomes `/32` or `/128`
-/// - CIDR keeps its configured prefix and is normalized to network address
-fn parse_persistent_ip_item(raw: &str, source: &str) -> Result<String> {
-    let value = raw.trim();
-    if value.is_empty() {
-        return Err(DnsError::plugin(format!("ros_route {source} is empty")));
-    }
-
-    if let Some((ip_raw, prefix_raw)) = value.split_once('/') {
-        let ip = ip_raw.trim().parse::<IpAddr>().map_err(|e| {
-            DnsError::plugin(format!("ros_route {source} has invalid ip '{ip_raw}': {e}"))
-        })?;
-        let prefix = prefix_raw.trim().parse::<u8>().map_err(|e| {
-            DnsError::plugin(format!(
-                "ros_route {source} has invalid prefix '{prefix_raw}': {e}"
-            ))
-        })?;
-        let max_prefix = if ip.is_ipv4() { 32 } else { 128 };
-        if prefix > max_prefix {
-            return Err(DnsError::plugin(format!(
-                "ros_route {source} has invalid prefix /{prefix} for {ip}, max /{max_prefix}"
-            )));
-        }
-        let network_ip = normalize_network_ip(ip, prefix);
-        return Ok(format!("{network_ip}/{prefix}"));
-    }
-
-    let ip = value.parse::<IpAddr>().map_err(|e| {
-        DnsError::plugin(format!("ros_route {source} has invalid ip '{value}': {e}"))
-    })?;
-    let prefix = if ip.is_ipv4() { 32 } else { 128 };
-    Ok(format!("{ip}/{prefix}"))
-}
-
-fn normalize_network_ip(ip: IpAddr, prefix: u8) -> IpAddr {
-    match ip {
-        IpAddr::V4(addr) => {
-            let raw = u32::from(addr);
-            let mask = if prefix == 0 {
-                0
-            } else {
-                u32::MAX << (32 - prefix)
-            };
-            IpAddr::V4(Ipv4Addr::from(raw & mask))
-        }
-        IpAddr::V6(addr) => {
-            let raw = u128::from(addr);
-            let mask = if prefix == 0 {
-                0
-            } else {
-                u128::MAX << (128 - prefix)
-            };
-            IpAddr::V6(Ipv6Addr::from(raw & mask))
-        }
-    }
-}
-
-#[inline]
-fn is_default_route_cidr(cidr: &str) -> bool {
-    cidr == "0.0.0.0/0" || cidr == "::/0"
-}
-
-/// Check whether this persistent route's family is enabled by gateway config.
-///
-/// Returns `Ok(false)` when family gateway is not configured so caller can skip
-/// the item without failing plugin startup.
-fn is_persistent_ip_family_enabled(
-    cidr: &str,
-    gateway4_enabled: bool,
-    gateway6_enabled: bool,
-    source: &str,
-) -> Result<bool> {
-    let (ip_raw, _) = cidr.split_once('/').ok_or_else(|| {
-        DnsError::plugin(format!(
-            "ros_route {source} has invalid normalized route '{cidr}'"
-        ))
-    })?;
-    let ip = ip_raw.parse::<IpAddr>().map_err(|e| {
-        DnsError::plugin(format!(
-            "ros_route {source} has invalid normalized route '{cidr}': {e}"
-        ))
-    })?;
-
-    match ip {
-        IpAddr::V4(_) if !gateway4_enabled => Ok(false),
-        IpAddr::V6(_) if !gateway6_enabled => Ok(false),
-        _ => Ok(true),
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use std::net::SocketAddr;
+    use std::net::{Ipv4Addr, Ipv6Addr, SocketAddr};
+    use std::time::Duration;
 
+    use serde_yaml_ng::Value;
+
+    use super::config::{DEFAULT_QUEUE_CAPACITY, DEFAULT_WAIT_TIMEOUT};
     use super::*;
     use crate::proto::rdata::{A, AAAA, CNAME, SOA};
     use crate::proto::{DNSClass, Message, Name, Question, RData, Record};
