+++ src/plugin/matcher/mod.rs
@@ -39,7 +39,7 @@ pub mod random;
 pub mod rate_limiter;
 pub mod rcode;
 pub mod resp_ip;
-pub mod rules;
+pub(crate) mod rules;
 pub mod string_exp;
 pub mod time;
 pub mod true_matcher;
