+++ src/plugin/registry/mod.rs
